@@ -1,6 +1,6 @@
 {
   "name": "react-starter",
-  "version": "0.7.0",
+  "version": "0.8.0",
   "description": "Koa react",
   "private": false,
   "main": "src/server/app.js",
@@ -31,15 +31,15 @@
     "generate-contributors": "all-contributors generate"
   },
   "dependencies": {
-    "@auxilin/common-logger": "0.5.0",
+    "@auxilin/common-logger": "0.5.1",
     "@hot-loader/react-dom": "16.9.0",
     "app-module-path": "2.2.0",
     "axios": "0.19.0",
     "classnames": "2.2.6",
     "connected-react-router": "6.5.2",
     "cssnano": "4.1.10",
-    "flow-bin": "0.105.2",
-    "handlebars": "4.1.2",
+    "flow-bin": "0.107.0",
+    "handlebars": "4.2.0",
     "history": "4.9.0",
     "html-webpack-plugin": "3.2.0",
     "incstr": "1.2.3",
@@ -48,48 +48,48 @@
     "koa-mount": "4.0.0",
     "koa-router": "7.4.0",
     "koa-static": "5.0.0",
-    "koa-views": "6.2.0",
+    "koa-views": "6.2.1",
     "koa-webpack": "5.2.4",
     "lodash": "4.17.15",
-    "nodemon": "1.19.1",
+    "nodemon": "1.19.2",
     "normalize.css": "8.0.1",
     "postcss-import": "12.0.1",
     "postcss-nested": "4.1.2",
     "postcss-preset-env": "6.7.0",
     "react": "16.9.0",
     "react-dom": "16.9.0",
-    "react-hot-loader": "4.12.11",
+    "react-hot-loader": "4.12.12",
     "react-icon-base": "2.1.2",
     "react-icons": "3.7.0",
-    "react-redux": "7.1.0",
+    "react-redux": "7.1.1",
     "react-router-dom": "5.0.1",
     "redux": "4.0.4",
     "redux-thunk": "2.3.0",
     "reselect": "4.0.0",
     "uuid": "3.3.3",
-    "webpack": "4.39.2",
-    "webpack-cli": "3.3.7",
+    "webpack": "4.39.3",
+    "webpack-cli": "3.3.8",
     "yup": "0.27.0"
   },
   "devDependencies": {
     "@auxilin/eslint-config": "0.2.0",
-    "@babel/cli": "7.5.5",
-    "@babel/core": "7.5.5",
+    "@babel/cli": "7.6.0",
+    "@babel/core": "7.6.0",
     "@babel/plugin-proposal-class-properties": "7.5.5",
     "@babel/plugin-proposal-export-namespace-from": "7.5.2",
     "@babel/plugin-syntax-dynamic-import": "7.2.0",
-    "@babel/preset-env": "7.5.5",
+    "@babel/preset-env": "7.6.0",
     "@babel/preset-flow": "7.0.0",
     "@babel/preset-react": "7.0.0",
-    "all-contributors-cli": "6.8.1",
-    "babel-eslint": "10.0.2",
+    "all-contributors-cli": "6.9.0",
+    "babel-eslint": "10.0.3",
     "babel-loader": "8.0.6",
     "babel-plugin-module-resolver": "3.2.0",
     "babel-plugin-react-css-modules": "5.2.6",
     "core-js": "3.2.1",
     "css-loader": "3.2.0",
-    "eslint": "6.2.0",
-    "eslint-plugin-flowtype": "4.2.0",
+    "eslint": "6.3.0",
+    "eslint-plugin-flowtype": "4.3.0",
     "eslint-plugin-import": "2.18.2",
     "eslint-plugin-jsx-a11y": "6.2.3",
     "eslint-plugin-react": "7.14.3",
