@@ -1,9 +1,9 @@
-// flow-typed signature: 00504217abb4effa078d1144a2bbba99
-// flow-typed version: 614bf49aa8/classnames_v2.x.x/flow_>=v0.25.x
+// flow-typed signature: 05f3dad8810e99e8737acdc3939d043d
+// flow-typed version: c6154227d1/classnames_v2.x.x/flow_>=v0.104.x
 
 type $npm$classnames$Classes =
   | string
-  | { [className: string]: * }
+  | { [className: string]: *, ... }
   | false
   | void
   | null;
