@@ -1,5 +1,5 @@
-// flow-typed signature: f60726754a93b34cc49e57e294b60644
-// flow-typed version: 0a8e7a7f8c/lodash_v4.x.x/flow_>=v0.63.x
+// flow-typed signature: 3584ccfacef49063d7c9a1743ce4ccad
+// flow-typed version: d381f02c43/lodash_v4.x.x/flow_>=v0.104.x
 
 declare module "lodash" {
   declare type Path = $ReadOnlyArray<string | number> | string | number;
@@ -140,34 +140,39 @@ declare module "lodash" {
     evaluate?: RegExp,
     imports?: Object,
     interpolate?: RegExp,
-    variable?: string
+    variable?: string,
+    ...
   };
 
   declare type TruncateOptions = {
     length?: number,
     omission?: string,
-    separator?: RegExp | string
+    separator?: RegExp | string,
+    ...
   };
 
   declare type Cancelable = {
     cancel: () => void,
-    flush: () => mixed
+    flush: () => mixed,
+    ...
   };
 
   declare type DebounceOptions = {
     leading?: boolean,
     maxWait?: number,
-    trailing?: boolean
+    trailing?: boolean,
+    ...
   };
 
   declare type ThrottleOptions = {
     leading?: boolean,
-    trailing?: boolean
+    trailing?: boolean,
+    ...
   };
 
   declare type NestedArray<T> = Array<Array<T>>;
 
-  declare type matchesIterateeShorthand = {[key: any]: any};
+  declare type matchesIterateeShorthand = { [key: any]: any, ... };
   declare type matchesPropertyIterateeShorthand = [string, any];
   declare type propertyIterateeShorthand = string;
 
@@ -218,7 +223,7 @@ declare module "lodash" {
 
   declare class Lodash {
     // Array
-    chunk<T>(array?: ?Array<T>, size?: ?number): Array<Array<T>>;
+    chunk<T>(array?: ?$ReadOnlyArray<T>, size?: ?number): Array<Array<T>>;
     compact<T, N: ?T>(array?: ?Array<N>): Array<T>;
     concat<T>(
       base?: ?$ReadOnlyArray<T>,
@@ -273,7 +278,7 @@ declare module "lodash" {
     flatten<T, X>(array?: ?$ReadOnlyArray<$ReadOnlyArray<T> | X>): Array<T | X>;
     flattenDeep<T>(array?: ?(any[])): Array<T>;
     flattenDepth(array?: ?(any[]), depth?: ?number): any[];
-    fromPairs<A, B>(pairs?: ?Array<[A, B]>): { [key: A]: B };
+    fromPairs<A, B>(pairs?: ?Array<[A, B]>): { [key: A]: B, ... };
     head<T>(array: ?$ReadOnlyArray<T>): T;
     indexOf<T>(array: Array<T>, value: T, fromIndex?: number): number;
     indexOf<T>(array: void | null, value?: ?T, fromIndex?: ?number): -1;
@@ -502,10 +507,10 @@ declare module "lodash" {
       a5: $ReadOnlyArray<E>
     ): Array<[A, B, C, D, E]>;
 
-    zipObject<K, V>(props: Array<K>, values?: ?Array<V>): { [key: K]: V };
-    zipObject<K, V>(props: void | null, values?: ?Array<V>): {};
+    zipObject<K, V>(props: Array<K>, values?: ?Array<V>): { [key: K]: V, ... };
+    zipObject<K, V>(props: void | null, values?: ?Array<V>): {...};
     zipObjectDeep(props: any[], values?: ?any): Object;
-    zipObjectDeep(props: void | null, values?: ?any): {};
+    zipObjectDeep(props: void | null, values?: ?any): {...};
 
     zipWith<A>(a1?: ?Array<A>): Array<[A]>;
     zipWith<T, A>(a1: Array<A>, iteratee: (A) => T): Array<T>;
@@ -545,7 +550,7 @@ declare module "lodash" {
 
     // Collection
     countBy<T>(array: Array<T>, iteratee?: ?ValueOnlyIteratee<T>): Object;
-    countBy<T>(array: void | null, iteratee?: ?ValueOnlyIteratee<T>): {};
+    countBy<T>(array: void | null, iteratee?: ?ValueOnlyIteratee<T>): {...};
     countBy<T: Object>(object: T, iteratee?: ?ValueOnlyIteratee<T>): Object;
     // alias of _.forEach
     each<T>(array: $ReadOnlyArray<T>, iteratee?: ?Iteratee<T>): Array<T>;
@@ -558,7 +563,7 @@ declare module "lodash" {
     every<T>(array?: ?$ReadOnlyArray<T>, iteratee?: ?Iteratee<T>): boolean;
     every<T: Object>(object: T, iteratee?: OIteratee<T>): boolean;
     filter<T>(array?: ?$ReadOnlyArray<T>, predicate?: ?Predicate<T>): Array<T>;
-    filter<A, T: { [id: any]: A }>(
+    filter<A, T: { [id: any]: A, ... }>(
       object: T,
       predicate?: OPredicate<A, T>
     ): Array<A>;
@@ -572,7 +577,7 @@ declare module "lodash" {
       predicate?: ?Predicate<T>,
       fromIndex?: ?number
     ): void;
-    find<V, A, T: { [id: any]: A }>(
+    find<V, A, T: { [id: any]: A, ... }>(
       object: T,
       predicate?: OPredicate<A, T>,
       fromIndex?: number
@@ -582,7 +587,7 @@ declare module "lodash" {
       predicate?: ?Predicate<T>,
       fromIndex?: ?number
     ): T | void;
-    findLast<V, A, T: { [id: any]: A }>(
+    findLast<V, A, T: { [id: any]: A, ... }>(
       object: T,
       predicate?: ?OPredicate<A, T>
     ): V;
@@ -624,12 +629,12 @@ declare module "lodash" {
     groupBy<V, T>(
       array: $ReadOnlyArray<T>,
       iteratee?: ?ValueOnlyIteratee<T>
-    ): { [key: V]: Array<T> };
-    groupBy(array: void | null, iteratee?: ?ValueOnlyIteratee<any>): {};
-    groupBy<V, A, T: { [id: any]: A }>(
+    ): { [key: V]: Array<T>, ... };
+    groupBy(array: void | null, iteratee?: ?ValueOnlyIteratee<any>): {...};
+    groupBy<V, A, T: { [id: any]: A, ... }>(
       object: T,
       iteratee?: ValueOnlyIteratee<A>
-    ): { [key: V]: Array<A> };
+    ): { [key: V]: Array<A>, ... };
     includes<T>(
       array: $ReadOnlyArray<T>,
       value: T,
@@ -651,12 +656,12 @@ declare module "lodash" {
     keyBy<T, V>(
       array: $ReadOnlyArray<T>,
       iteratee?: ?ValueOnlyIteratee<T>
-    ): { [key: V]: T };
-    keyBy(array: void | null, iteratee?: ?ValueOnlyIteratee<*>): {};
-    keyBy<V, A, I, T: { [id: I]: A }>(
+    ): { [key: V]: T, ... };
+    keyBy(array: void | null, iteratee?: ?ValueOnlyIteratee<*>): {...};
+    keyBy<V, A, I, T: { [id: I]: A, ... }>(
       object: T,
       iteratee?: ?ValueOnlyIteratee<A>
-    ): { [key: V]: A };
+    ): { [key: V]: A, ... };
     map<T, U>(array?: ?Array<T>, iteratee?: ?MapIterator<T, U>): Array<U>;
     map<T, U>(
       array: ?$ReadOnlyArray<T>,
@@ -680,7 +685,7 @@ declare module "lodash" {
       iteratees?: ?$ReadOnlyArray<Iteratee<T>> | ?string,
       orders?: ?$ReadOnlyArray<"asc" | "desc"> | ?string
     ): Array<T>;
-    orderBy<V, T: {}>(
+    orderBy<V, T: {...}>(
       object: T,
       iteratees?: $ReadOnlyArray<OIteratee<*>> | string,
       orders?: $ReadOnlyArray<"asc" | "desc"> | string
@@ -689,7 +694,7 @@ declare module "lodash" {
       array?: ?$ReadOnlyArray<T>,
       predicate?: ?Predicate<T>
     ): [Array<T>, Array<T>];
-    partition<V, A, T: { [id: any]: A }>(
+    partition<V, A, T: { [id: any]: A, ... }>(
       object: T,
       predicate?: OPredicate<A, T>
     ): [Array<V>, Array<V>];
@@ -744,7 +749,7 @@ declare module "lodash" {
       accumulator?: ?U
     ): U;
     reject<T>(array: ?$ReadOnlyArray<T>, predicate?: Predicate<T>): Array<T>;
-    reject<V: Object, A, T: { [id: any]: A }>(
+    reject<V: Object, A, T: { [id: any]: A, ... }>(
       object?: ?T,
       predicate?: ?OPredicate<A, T>
     ): Array<V>;
@@ -757,7 +762,7 @@ declare module "lodash" {
     size(collection: $ReadOnlyArray<any> | Object | string): number;
     some<T>(array: void | null, predicate?: ?Predicate<T>): false;
     some<T>(array: ?$ReadOnlyArray<T>, predicate?: Predicate<T>): boolean;
-    some<A, T: { [id: any]: A }>(
+    some<A, T: { [id: any]: A, ... }>(
       object?: ?T,
       predicate?: OPredicate<A, T>
     ): boolean;
@@ -830,9 +835,9 @@ declare module "lodash" {
       value: T,
       customizer?: ?(value: T, key: number | string, object: T, stack: any) => U
     ): U;
-    conformsTo<T: { [key: string]: mixed }>(
+    conformsTo<T: { [key: string]: mixed, ... }>(
       source: T,
-      predicates: T & { [key: string]: (x: any) => boolean }
+      predicates: T & { [key: string]: (x: any) => boolean, ... }
     ): boolean;
     eq(value: any, other: any): boolean;
     gt(value: any, other: any): boolean;
@@ -843,9 +848,9 @@ declare module "lodash" {
     isArray(value: any): false;
     isArrayBuffer(value: ArrayBuffer): true;
     isArrayBuffer(value: any): false;
-    isArrayLike(value: Array<any> | string | { length: number }): true;
+    isArrayLike(value: Array<any> | string | { length: number, ... }): true;
     isArrayLike(value: any): false;
-    isArrayLikeObject(value: { length: number } | Array<any>): true;
+    isArrayLikeObject(value: { length: number, ... } | Array<any>): true;
     isArrayLikeObject(value: any): false;
     isBoolean(value: boolean): true;
     isBoolean(value: any): false;
@@ -855,7 +860,7 @@ declare module "lodash" {
     isDate(value: any): false;
     isElement(value: Element): true;
     isElement(value: any): false;
-    isEmpty(value: void | null | "" | {} | [] | number | boolean): true;
+    isEmpty(value: void | null | "" | {...} | [] | number | boolean): true;
     isEmpty(value: any): boolean;
     isEqual(value: any, other: any): boolean;
     isEqualWith<T, U>(
@@ -968,12 +973,12 @@ declare module "lodash" {
 
     // Object
     assign(object?: ?Object, ...sources?: Array<?Object>): Object;
-    assignIn(): {};
+    assignIn(): {...};
     assignIn<A, B>(a: A, b: B): A & B;
     assignIn<A, B, C>(a: A, b: B, c: C): A & B & C;
     assignIn<A, B, C, D>(a: A, b: B, c: C, d: D): A & B & C & D;
     assignIn<A, B, C, D, E>(a: A, b: B, c: C, d: D, e: E): A & B & C & D & E;
-    assignInWith(): {};
+    assignInWith(): {...};
     assignInWith<T: Object, A: Object>(
       object: T,
       s1: A,
@@ -1024,7 +1029,7 @@ declare module "lodash" {
         source: A | B | C | D
       ) => any | void
     ): Object;
-    assignWith(): {};
+    assignWith(): {...};
     assignWith<T: Object, A: Object>(
       object: T,
       s1: A,
@@ -1077,9 +1082,9 @@ declare module "lodash" {
     ): Object;
     at(object?: ?Object, ...paths: Array<string>): Array<any>;
     at(object?: ?Object, paths: Array<string>): Array<any>;
-    create(prototype: void | null, properties: void | null): {};
+    create(prototype: void | null, properties: void | null): {...};
     create<T>(prototype: T, properties: Object): T;
-    create(prototype: any, properties: void | null): {};
+    create(prototype: any, properties: void | null): {...};
     defaults(object?: ?Object, ...sources?: Array<?Object>): Object;
     defaultsDeep(object?: ?Object, ...sources?: Array<?Object>): Object;
     // alias for _.toPairs
@@ -1142,19 +1147,19 @@ declare module "lodash" {
         source: A | B | C | D
       ) => any | void
     ): Object;
-    findKey<A, T: { [id: any]: A }>(
+    findKey<A, T: { [id: any]: A, ... }>(
       object: T,
       predicate?: ?OPredicate<A, T>
     ): string | void;
-    findKey<A, T: { [id: any]: A }>(
+    findKey<A, T: { [id: any]: A, ... }>(
       object: void | null,
       predicate?: ?OPredicate<A, T>
     ): void;
-    findLastKey<A, T: { [id: any]: A }>(
+    findLastKey<A, T: { [id: any]: A, ... }>(
       object: T,
       predicate?: ?OPredicate<A, T>
     ): string | void;
-    findLastKey<A, T: { [id: any]: A }>(
+    findLastKey<A, T: { [id: any]: A, ... }>(
       object: void | null,
       predicate?: ?OPredicate<A, T>
     ): void;
@@ -1180,23 +1185,23 @@ declare module "lodash" {
     hasIn(object: Object, path: void | null): false;
     hasIn(object: void | null, path?: ?Path): false;
     invert(object: Object, multiVal?: ?boolean): Object;
-    invert(object: void | null, multiVal?: ?boolean): {};
+    invert(object: void | null, multiVal?: ?boolean): {...};
     invertBy(object: Object, iteratee?: ?Function): Object;
-    invertBy(object: void | null, iteratee?: ?Function): {};
+    invertBy(object: void | null, iteratee?: ?Function): {...};
     invoke(
       object?: ?Object,
       path?: ?Path,
       ...args?: Array<any>
     ): any;
-    keys<K>(object?: ?{ [key: K]: any }): Array<K>;
+    keys<K>(object?: ?{ [key: K]: any, ... }): Array<K>;
     keys(object?: ?Object): Array<string>;
     keysIn(object?: ?Object): Array<string>;
     mapKeys(object: Object, iteratee?: ?OIteratee<*>): Object;
-    mapKeys(object: void | null, iteratee?: ?OIteratee<*>): {};
+    mapKeys(object: void | null, iteratee?: ?OIteratee<*>): {...};
     mapValues(object: Object, iteratee?: ?OIteratee<*>): Object;
-    mapValues(object: void | null, iteratee?: ?OIteratee<*>): {};
+    mapValues(object: void | null, iteratee?: ?OIteratee<*>): {...};
     merge(object?: ?Object, ...sources?: Array<?Object>): Object;
-    mergeWith(): {};
+    mergeWith(): {...};
     mergeWith<T: Object, A: Object>(
       object: T,
       customizer?: (
@@ -1248,18 +1253,18 @@ declare module "lodash" {
     ): Object;
     omit(object?: ?Object, ...props: Array<string>): Object;
     omit(object?: ?Object, props: Array<string>): Object;
-    omitBy<A, T: { [id: any]: A } | { [id: number]: A }>(
+    omitBy<A, T: { [id: any]: A, ... } | { [id: number]: A, ... }>(
       object: T,
       predicate?: ?OPredicate<A, T>
     ): Object;
-    omitBy<A, T>(object: void | null, predicate?: ?OPredicate<A, T>): {};
+    omitBy<A, T>(object: void | null, predicate?: ?OPredicate<A, T>): {...};
     pick(object?: ?Object, ...props: Array<string>): Object;
     pick(object?: ?Object, props: $ReadOnlyArray<string>): Object;
-    pickBy<A, T: { [id: any]: A } | { [id: number]: A }>(
+    pickBy<A, T: { [id: any]: A, ... } | { [id: number]: A, ... }>(
       object: T,
       predicate?: ?OPredicate<A, T>
     ): Object;
-    pickBy<A, T>(object: void | null, predicate?: ?OPredicate<A, T>): {};
+    pickBy<A, T>(object: void | null, predicate?: ?OPredicate<A, T>): {...};
     result(
       object?: ?Object,
       path?: ?Path,
@@ -1294,7 +1299,7 @@ declare module "lodash" {
       collection: void | null,
       iteratee?: ?OIteratee<*>,
       accumulator?: ?any
-    ): {};
+    ): {...};
     unset(object: void | null, path?: ?Path): true;
     unset(object: Object, path?: ?Path): boolean;
     update(object: Object, path: Path, updater: Function): Object;
@@ -1423,7 +1428,7 @@ declare module "lodash" {
     mixin<T: Function | Object>(
       object?: T,
       source: Object,
-      options?: { chain: boolean }
+      options?: { chain: boolean, ... }
     ): T;
     noConflict(): Lodash;
     noop(...args: Array<mixed>): void;
@@ -1444,7 +1449,7 @@ declare module "lodash" {
 
     stubArray(): Array<*>;
     stubFalse(): false;
-    stubObject(): {};
+    stubObject(): {...};
     stubString(): "";
     stubTrue(): true;
     times(n?: ?number, ...rest?: Array<void | null>): Array<number>;
@@ -1599,29 +1604,33 @@ declare module "lodash/fp" {
     evaluate?: RegExp,
     imports?: Object,
     interpolate?: RegExp,
-    variable?: string
+    variable?: string,
+    ...
   };
 
   declare type TruncateOptions = {
     length?: number,
     omission?: string,
-    separator?: RegExp | string
+    separator?: RegExp | string,
+    ...
   };
 
   declare type DebounceOptions = {
     leading?: boolean,
     maxWait?: number,
-    trailing?: boolean
+    trailing?: boolean,
+    ...
   };
 
   declare type ThrottleOptions = {
     leading?: boolean,
-    trailing?: boolean
+    trailing?: boolean,
+    ...
   };
 
   declare type NestedArray<T> = Array<Array<T>>;
 
-  declare type matchesIterateeShorthand = {[string | number]: any};
+  declare type matchesIterateeShorthand = { [string | number]: any, ... };
   declare type matchesPropertyIterateeShorthand = [string, any];
   declare type propertyIterateeShorthand = string;
 
@@ -1772,7 +1781,7 @@ declare module "lodash/fp" {
     flattenDeep<T>(array: any[]): Array<T>;
     flattenDepth(depth: number): (array: any[]) => any[];
     flattenDepth(depth: number, array: any[]): any[];
-    fromPairs<A, B>(pairs: Array<[A, B]>): { [key: A]: B };
+    fromPairs<A, B>(pairs: Array<[A, B]>): { [key: A]: B, ... };
     head<T>(array: $ReadOnlyArray<T>): T;
     indexOf<T>(value: T): (array: Array<T>) => number;
     indexOf<T>(value: T, array: Array<T>): number;
@@ -2004,8 +2013,8 @@ declare module "lodash/fp" {
     zip<A, B>(a1: A[]): (a2: B[]) => Array<[A, B]>;
     zip<A, B>(a1: A[], a2: B[]): Array<[A, B]>;
     zipAll(arrays: Array<Array<any>>): Array<any>;
-    zipObject<K, V>(props?: Array<K>): (values?: Array<V>) => { [key: K]: V };
-    zipObject<K, V>(props?: Array<K>, values?: Array<V>): { [key: K]: V };
+    zipObject<K, V>(props?: Array<K>): (values?: Array<V>) => { [key: K]: V, ... };
+    zipObject<K, V>(props?: Array<K>, values?: Array<V>): { [key: K]: V, ... };
     zipObj(props: Array<any>): (values: Array<any>) => Object;
     zipObj(props: Array<any>, values: Array<any>): Object;
     zipObjectDeep(props: any[]): (values: any) => Object;
@@ -2026,158 +2035,158 @@ declare module "lodash/fp" {
     // Collection
     countBy<T>(
       iteratee: ValueOnlyIteratee<T>
-    ): (collection: Array<T> | { [id: any]: T }) => { [string]: number };
+    ): (collection: Array<T> | { [id: any]: T, ... }) => { [string]: number, ... };
     countBy<T>(
       iteratee: ValueOnlyIteratee<T>,
-      collection: Array<T> | { [id: any]: T }
-    ): { [string]: number };
+      collection: Array<T> | { [id: any]: T, ... }
+    ): { [string]: number, ... };
     // alias of _.forEach
     each<T>(
       iteratee: Iteratee<T> | OIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>;
     each<T>(
       iteratee: Iteratee<T> | OIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<T>;
     // alias of _.forEachRight
     eachRight<T>(
       iteratee: Iteratee<T> | OIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>;
     eachRight<T>(
       iteratee: Iteratee<T> | OIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<T>;
     every<T>(
       iteratee: Iteratee<T> | OIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => boolean;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => boolean;
     every<T>(
       iteratee: Iteratee<T> | OIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): boolean;
     all<T>(
       iteratee: Iteratee<T> | OIteratee<T>
-    ): (collection: Array<T> | { [id: any]: T }) => boolean;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => boolean;
     all<T>(
       iteratee: Iteratee<T> | OIteratee<T>,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): boolean;
     filter<T>(
       predicate: Predicate<T> | OPredicate<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>;
     filter<T>(
       predicate: Predicate<T> | OPredicate<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<T>;
     find<T>(
       predicate: Predicate<T> | OPredicate<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => T | void;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => T | void;
     find<T>(
       predicate: Predicate<T> | OPredicate<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): T | void;
     findFrom<T>(
       predicate: Predicate<T> | OPredicate<T>
     ): ((
       fromIndex: number
-    ) => (collection: $ReadOnlyArray<T> | { [id: any]: T }) => T | void) &
+    ) => (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => T | void) &
       ((
         fromIndex: number,
-        collection: $ReadOnlyArray<T> | { [id: any]: T }
+        collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
       ) => T | void);
     findFrom<T>(
       predicate: Predicate<T> | OPredicate<T>,
       fromIndex: number
-    ): (collection: Array<T> | { [id: any]: T }) => T | void;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => T | void;
     findFrom<T>(
       predicate: Predicate<T> | OPredicate<T>,
       fromIndex: number,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): T | void;
     findLast<T>(
       predicate: Predicate<T> | OPredicate<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => T | void;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => T | void;
     findLast<T>(
       predicate: Predicate<T> | OPredicate<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): T | void;
     findLastFrom<T>(
       predicate: Predicate<T> | OPredicate<T>
     ): ((
       fromIndex: number
-    ) => (collection: $ReadOnlyArray<T> | { [id: any]: T }) => T | void) &
+    ) => (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => T | void) &
       ((
         fromIndex: number,
-        collection: $ReadOnlyArray<T> | { [id: any]: T }
+        collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
       ) => T | void);
     findLastFrom<T>(
       predicate: Predicate<T> | OPredicate<T>,
       fromIndex: number
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => T | void;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => T | void;
     findLastFrom<T>(
       predicate: Predicate<T> | OPredicate<T>,
       fromIndex: number,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): T | void;
     flatMap<T, U>(
       iteratee: FlatMapIteratee<T, U> | OFlatMapIteratee<T, U>
-    ): (collection: Array<T> | { [id: any]: T }) => Array<U>;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => Array<U>;
     flatMap<T, U>(
       iteratee: FlatMapIteratee<T, U> | OFlatMapIteratee<T, U>,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): Array<U>;
     flatMapDeep<T, U>(
       iteratee: FlatMapIteratee<T, U> | OFlatMapIteratee<T, U>
-    ): (collection: Array<T> | { [id: any]: T }) => Array<U>;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => Array<U>;
     flatMapDeep<T, U>(
       iteratee: FlatMapIteratee<T, U> | OFlatMapIteratee<T, U>,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): Array<U>;
     flatMapDepth<T, U>(
       iteratee: FlatMapIteratee<T, U> | OFlatMapIteratee<T, U>
     ): ((
       depth: number
-    ) => (collection: Array<T> | { [id: any]: T }) => Array<U>) &
-      ((depth: number, collection: Array<T> | { [id: any]: T }) => Array<U>);
+    ) => (collection: Array<T> | { [id: any]: T, ... }) => Array<U>) &
+      ((depth: number, collection: Array<T> | { [id: any]: T, ... }) => Array<U>);
     flatMapDepth<T, U>(
       iteratee: FlatMapIteratee<T, U> | OFlatMapIteratee<T, U>,
       depth: number
-    ): (collection: Array<T> | { [id: any]: T }) => Array<U>;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => Array<U>;
     flatMapDepth<T, U>(
       iteratee: FlatMapIteratee<T, U> | OFlatMapIteratee<T, U>,
       depth: number,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): Array<U>;
     forEach<T>(
       iteratee: Iteratee<T> | OIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>;
     forEach<T>(
       iteratee: Iteratee<T> | OIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<T>;
     forEachRight<T>(
       iteratee: Iteratee<T> | OIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>;
     forEachRight<T>(
       iteratee: Iteratee<T> | OIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<T>;
     groupBy<V, T>(
       iteratee: ValueOnlyIteratee<T>
     ): (
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
-    ) => { [key: V]: Array<T> };
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
+    ) => { [key: V]: Array<T>, ... };
     groupBy<V, T>(
       iteratee: ValueOnlyIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
-    ): { [key: V]: Array<T> };
-    includes<T>(value: T): (collection: Array<T> | { [id: any]: T }) => boolean;
-    includes<T>(value: T, collection: Array<T> | { [id: any]: T }): boolean;
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
+    ): { [key: V]: Array<T>, ... };
+    includes<T>(value: T): (collection: Array<T> | { [id: any]: T, ... }) => boolean;
+    includes<T>(value: T, collection: Array<T> | { [id: any]: T, ... }): boolean;
     includes(value: string): (str: string) => boolean;
     includes(value: string, str: string): boolean;
     contains(value: string): (str: string) => boolean;
     contains(value: string, str: string): boolean;
-    contains<T>(value: T): (collection: Array<T> | { [id: any]: T }) => boolean;
-    contains<T>(value: T, collection: Array<T> | { [id: any]: T }): boolean;
+    contains<T>(value: T): (collection: Array<T> | { [id: any]: T, ... }) => boolean;
+    contains<T>(value: T, collection: Array<T> | { [id: any]: T, ... }): boolean;
     includesFrom(
       value: string
     ): ((fromIndex: number) => (str: string) => boolean) &
@@ -2195,58 +2204,58 @@ declare module "lodash/fp" {
     includesFrom<T>(value: T, fromIndex: number, collection: Array<T>): boolean;
     invokeMap<T>(
       path: ((value: T) => Path) | Path
-    ): (collection: Array<T> | { [id: any]: T }) => Array<any>;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => Array<any>;
     invokeMap<T>(
       path: ((value: T) => Path) | Path,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): Array<any>;
     invokeArgsMap<T>(
       path: ((value: T) => Path) | Path
     ): ((
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ) => (args: Array<any>) => Array<any>) &
       ((
-        collection: Array<T> | { [id: any]: T },
+        collection: Array<T> | { [id: any]: T, ... },
         args: Array<any>
       ) => Array<any>);
     invokeArgsMap<T>(
       path: ((value: T) => Path) | Path,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): (args: Array<any>) => Array<any>;
     invokeArgsMap<T>(
       path: ((value: T) => Path) | Path,
-      collection: Array<T> | { [id: any]: T },
+      collection: Array<T> | { [id: any]: T, ... },
       args: Array<any>
     ): Array<any>;
     keyBy<T, V>(
       iteratee: ValueOnlyIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => { [key: V]: T };
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => { [key: V]: T, ... };
     keyBy<T, V>(
       iteratee: ValueOnlyIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
-    ): { [key: V]: T };
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
+    ): { [key: V]: T, ... };
     indexBy<T, V>(
       iteratee: ValueOnlyIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => { [key: V]: T };
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => { [key: V]: T, ... };
     indexBy<T, V>(
       iteratee: ValueOnlyIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
-    ): { [key: V]: T };
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
+    ): { [key: V]: T, ... };
     map<T, U>(
       iteratee: MapIterator<T, U> | OMapIterator<T, U>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<U>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<U>;
     map<T, U>(
       iteratee: MapIterator<T, U> | OMapIterator<T, U>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<U>;
     map(iteratee: (char: string) => any): (str: string) => string;
     map(iteratee: (char: string) => any, str: string): string;
     pluck<T, U>(
       iteratee: MapIterator<T, U> | OMapIterator<T, U>
-    ): (collection: Array<T> | { [id: any]: T }) => Array<U>;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => Array<U>;
     pluck<T, U>(
       iteratee: MapIterator<T, U> | OMapIterator<T, U>,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): Array<U>;
     pluck(iteratee: (char: string) => any): (str: string) => string;
     pluck(iteratee: (char: string) => any, str: string): string;
@@ -2254,93 +2263,93 @@ declare module "lodash/fp" {
       iteratees: $ReadOnlyArray<Iteratee<T> | OIteratee<*>> | string
     ): ((
       orders: $ReadOnlyArray<"asc" | "desc"> | string
-    ) => (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>) &
+    ) => (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>) &
       ((
         orders: $ReadOnlyArray<"asc" | "desc"> | string,
-        collection: $ReadOnlyArray<T> | { [id: any]: T }
+        collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
       ) => Array<T>);
     orderBy<T>(
       iteratees: $ReadOnlyArray<Iteratee<T> | OIteratee<*>> | string,
       orders: $ReadOnlyArray<"asc" | "desc"> | string
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>;
     orderBy<T>(
       iteratees: $ReadOnlyArray<Iteratee<T> | OIteratee<*>> | string,
       orders: $ReadOnlyArray<"asc" | "desc"> | string,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<T>;
     partition<T>(
       predicate: Predicate<T> | OPredicate<T>
-    ): (collection: Array<T> | { [id: any]: T }) => [Array<T>, Array<T>];
+    ): (collection: Array<T> | { [id: any]: T, ... }) => [Array<T>, Array<T>];
     partition<T>(
       predicate: Predicate<T> | OPredicate<T>,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): [Array<T>, Array<T>];
     reduce<T, U>(
       iteratee: (accumulator: U, value: T) => U
-    ): ((accumulator: U) => (collection: Array<T> | { [id: any]: T }) => U) &
-      ((accumulator: U, collection: Array<T> | { [id: any]: T }) => U);
+    ): ((accumulator: U) => (collection: Array<T> | { [id: any]: T, ... }) => U) &
+      ((accumulator: U, collection: Array<T> | { [id: any]: T, ... }) => U);
     reduce<T, U>(
       iteratee: (accumulator: U, value: T) => U,
       accumulator: U
-    ): (collection: Array<T> | { [id: any]: T }) => U;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => U;
     reduce<T, U>(
       iteratee: (accumulator: U, value: T) => U,
       accumulator: U,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): U;
     reduceRight<T, U>(
       iteratee: (value: T, accumulator: U) => U
-    ): ((accumulator: U) => (collection: Array<T> | { [id: any]: T }) => U) &
-      ((accumulator: U, collection: Array<T> | { [id: any]: T }) => U);
+    ): ((accumulator: U) => (collection: Array<T> | { [id: any]: T, ... }) => U) &
+      ((accumulator: U, collection: Array<T> | { [id: any]: T, ... }) => U);
     reduceRight<T, U>(
       iteratee: (value: T, accumulator: U) => U,
       accumulator: U
-    ): (collection: Array<T> | { [id: any]: T }) => U;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => U;
     reduceRight<T, U>(
       iteratee: (value: T, accumulator: U) => U,
       accumulator: U,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): U;
     reject<T>(
       predicate: Predicate<T> | OPredicate<T>
-    ): (collection: Array<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => Array<T>;
     reject<T>(
       predicate: Predicate<T> | OPredicate<T>,
-      collection: Array<T> | { [id: any]: T }
+      collection: Array<T> | { [id: any]: T, ... }
     ): Array<T>;
-    sample<T>(collection: Array<T> | { [id: any]: T }): T;
+    sample<T>(collection: Array<T> | { [id: any]: T, ... }): T;
     sampleSize<T>(
       n: number
-    ): (collection: Array<T> | { [id: any]: T }) => Array<T>;
-    sampleSize<T>(n: number, collection: Array<T> | { [id: any]: T }): Array<T>;
-    shuffle<T>(collection: Array<T> | { [id: any]: T }): Array<T>;
+    ): (collection: Array<T> | { [id: any]: T, ... }) => Array<T>;
+    sampleSize<T>(n: number, collection: Array<T> | { [id: any]: T, ... }): Array<T>;
+    shuffle<T>(collection: Array<T> | { [id: any]: T, ... }): Array<T>;
     size(collection: $ReadOnlyArray<any> | Object | string): number;
     some<T>(
       predicate: Predicate<T> | OPredicate<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => boolean;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => boolean;
     some<T>(
       predicate: Predicate<T> | OPredicate<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): boolean;
     any<T>(
       predicate: Predicate<T> | OPredicate<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => boolean;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => boolean;
     any<T>(
       predicate: Predicate<T> | OPredicate<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): boolean;
     sortBy<T>(
       iteratees:
         | $ReadOnlyArray<Iteratee<T> | OIteratee<T>>
         | Iteratee<T>
         | OIteratee<T>
-    ): (collection: $ReadOnlyArray<T> | { [id: any]: T }) => Array<T>;
+    ): (collection: $ReadOnlyArray<T> | { [id: any]: T, ... }) => Array<T>;
     sortBy<T>(
       iteratees:
         | $ReadOnlyArray<Iteratee<T> | OIteratee<T>>
         | Iteratee<T>
         | OIteratee<T>,
-      collection: $ReadOnlyArray<T> | { [id: any]: T }
+      collection: $ReadOnlyArray<T> | { [id: any]: T, ... }
     ): Array<T>;
 
     // Date
@@ -2416,25 +2425,25 @@ declare module "lodash/fp" {
       customizer: (value: T, key: number | string, object: T, stack: any) => U,
       value: T
     ): U;
-    conformsTo<T: { [key: string]: mixed }>(
-      predicates: T & { [key: string]: (x: any) => boolean }
+    conformsTo<T: { [key: string]: mixed, ... }>(
+      predicates: T & { [key: string]: (x: any) => boolean, ... }
     ): (source: T) => boolean;
-    conformsTo<T: { [key: string]: mixed }>(
-      predicates: T & { [key: string]: (x: any) => boolean },
+    conformsTo<T: { [key: string]: mixed, ... }>(
+      predicates: T & { [key: string]: (x: any) => boolean, ... },
       source: T
     ): boolean;
-    where<T: { [key: string]: mixed }>(
-      predicates: T & { [key: string]: (x: any) => boolean }
+    where<T: { [key: string]: mixed, ... }>(
+      predicates: T & { [key: string]: (x: any) => boolean, ... }
     ): (source: T) => boolean;
-    where<T: { [key: string]: mixed }>(
-      predicates: T & { [key: string]: (x: any) => boolean },
+    where<T: { [key: string]: mixed, ... }>(
+      predicates: T & { [key: string]: (x: any) => boolean, ... },
       source: T
     ): boolean;
-    conforms<T: { [key: string]: mixed }>(
-      predicates: T & { [key: string]: (x: any) => boolean }
+    conforms<T: { [key: string]: mixed, ... }>(
+      predicates: T & { [key: string]: (x: any) => boolean, ... }
     ): (source: T) => boolean;
-    conforms<T: { [key: string]: mixed }>(
-      predicates: T & { [key: string]: (x: any) => boolean },
+    conforms<T: { [key: string]: mixed, ... }>(
+      predicates: T & { [key: string]: (x: any) => boolean, ... },
       source: T
     ): boolean;
     eq(value: any): (other: any) => boolean;
@@ -2782,17 +2791,17 @@ declare module "lodash/fp" {
       object: T,
       s1: A
     ): Object;
-    findKey<A, T: { [id: any]: A }>(
+    findKey<A, T: { [id: any]: A, ... }>(
       predicate: OPredicate<A>
     ): (object: T) => string | void;
-    findKey<A, T: { [id: any]: A }>(
+    findKey<A, T: { [id: any]: A, ... }>(
       predicate: OPredicate<A>,
       object: T
     ): string | void;
-    findLastKey<A, T: { [id: any]: A }>(
+    findLastKey<A, T: { [id: any]: A, ... }>(
       predicate: OPredicate<A>
     ): (object: T) => string | void;
-    findLastKey<A, T: { [id: any]: A }>(
+    findLastKey<A, T: { [id: any]: A, ... }>(
       predicate: OPredicate<A>,
       object: T
     ): string | void;
@@ -2888,7 +2897,7 @@ declare module "lodash/fp" {
       object: Object,
       args: Array<any>
     ): any;
-    keys<K>(object: { [key: K]: any }): Array<K>;
+    keys<K>(object: { [key: K]: any, ... }): Array<K>;
     keys(object: Object): Array<string>;
     keysIn(object: Object): Array<string>;
     mapKeys(iteratee: OIteratee<*>): (object: Object) => Object;
@@ -2951,20 +2960,20 @@ declare module "lodash/fp" {
     omit(props: Array<string>, object: Object): Object;
     omitAll(props: Array<string>): (object: Object) => Object;
     omitAll(props: Array<string>, object: Object): Object;
-    omitBy<A, T: { [id: any]: A }>(
+    omitBy<A, T: { [id: any]: A, ... }>(
       predicate: OPredicate<A>
     ): (object: T) => Object;
-    omitBy<A, T: { [id: any]: A }>(predicate: OPredicate<A>, object: T): Object;
-    pick(...props: Array<string | {}>): Object;
+    omitBy<A, T: { [id: any]: A, ... }>(predicate: OPredicate<A>, object: T): Object;
+    pick(...props: Array<string | {...}>): Object;
     pick(props: $ReadOnlyArray<string>, object: Object): Object;
     pick(...props: Array<string>): (object: Object) => Object;
     pick(props: $ReadOnlyArray<string>): (object: Object) => Object;
     pickAll(props: Array<string>): (object: Object) => Object;
     pickAll(props: Array<string>, object: Object): Object;
-    pickBy<A, T: { [id: any]: A }>(
+    pickBy<A, T: { [id: any]: A, ... }>(
       predicate: OPredicate<A>
     ): (object: T) => Object;
-    pickBy<A, T: { [id: any]: A }>(predicate: OPredicate<A>, object: T): Object;
+    pickBy<A, T: { [id: any]: A, ... }>(predicate: OPredicate<A>, object: T): Object;
     result(path: Path): (object: Object) => any;
     result(path: Path, object: Object): any;
     set(
@@ -3199,16 +3208,16 @@ declare module "lodash/fp" {
     methodOf(object: Object): Function;
     mixin<T: Function | Object>(
       object: T
-    ): ((source: Object) => (options: { chain: boolean }) => T) &
-      ((source: Object, options: { chain: boolean }) => T);
+    ): ((source: Object) => (options: { chain: boolean, ... }) => T) &
+      ((source: Object, options: { chain: boolean, ... }) => T);
     mixin<T: Function | Object>(
       object: T,
       source: Object
-    ): (options: { chain: boolean }) => T;
+    ): (options: { chain: boolean, ... }) => T;
     mixin<T: Function | Object>(
       object: T,
       source: Object,
-      options: { chain: boolean }
+      options: { chain: boolean, ... }
     ): T;
     noConflict(): Lodash;
     noop(...args: Array<mixed>): void;
@@ -3246,7 +3255,7 @@ declare module "lodash/fp" {
     stubArray(): Array<*>;
     stubFalse(): false;
     F(): false;
-    stubObject(): {};
+    stubObject(): {...};
     stubString(): "";
     stubTrue(): true;
     T(): true;
@@ -3263,7 +3272,8 @@ declare module "lodash/fp" {
       curry?: boolean,
       fixed?: boolean,
       immutable?: boolean,
-      rearg?: boolean
+      rearg?: boolean,
+      ...
     }): void;
 
     // Properties
