@@ -1,5 +1,5 @@
-// flow-typed signature: e417325c3649446aca084d6e81845b5a
-// flow-typed version: 2973a15489/react-router-dom_v5.x.x/flow_>=v0.98.x
+// flow-typed signature: 0ebce0f11f06f21398113e8b4f68fa07
+// flow-typed version: c6154227d1/react-router-dom_v5.x.x/flow_>=v0.104.x
 
 declare module "react-router-dom" {
   declare export var BrowserRouter: React$ComponentType<{|
@@ -21,19 +21,21 @@ declare module "react-router-dom" {
     className?: string,
     to: string | LocationShape,
     replace?: boolean,
-    children?: React$Node
+    children?: React$Node,
+    ...
   }>
 
   declare export var NavLink: React$ComponentType<{
     to: string | LocationShape,
     activeClassName?: string,
     className?: string,
-    activeStyle?: { +[string]: mixed },
-    style?: { +[string]: mixed },
+    activeStyle?: { +[string]: mixed, ... },
+    style?: { +[string]: mixed, ... },
     isActive?: (match: Match, location: Location) => boolean,
     children?: React$Node,
     exact?: boolean,
-    strict?: boolean
+    strict?: boolean,
+    ...
   }>
 
   // NOTE: Below are duplicated from react-router. If updating these, please
@@ -43,14 +45,16 @@ declare module "react-router-dom" {
     search: string,
     hash: string,
     state?: any,
-    key?: string
+    key?: string,
+    ...
   };
 
   declare export type LocationShape = {
     pathname?: string,
     search?: string,
     hash?: string,
-    state?: any
+    state?: any,
+    ...
   };
 
   declare export type HistoryAction = "PUSH" | "REPLACE" | "POP";
@@ -73,14 +77,16 @@ declare module "react-router-dom" {
     ): () => void,
     // createMemoryHistory
     index?: number,
-    entries?: Array<Location>
+    entries?: Array<Location>,
+    ...
   };
 
   declare export type Match = {
-    params: { [key: string]: ?string },
+    params: { [key: string]: ?string, ... },
     isExact: boolean,
     path: string,
-    url: string
+    url: string,
+    ...
   };
 
   declare export type ContextRouter = {|
@@ -94,7 +100,8 @@ declare module "react-router-dom" {
     history: RouterHistory | void,
     location: Location | void,
     match: Match | void,
-    staticContext?: StaticRouterContext | void
+    staticContext?: StaticRouterContext | void,
+    ...
   };
 
   declare export type GetUserConfirmation = (
@@ -102,9 +109,7 @@ declare module "react-router-dom" {
     callback: (confirmed: boolean) => void
   ) => void;
 
-  declare export type StaticRouterContext = {
-    url?: string
-  };
+  declare export type StaticRouterContext = { url?: string, ... };
 
   declare export var StaticRouter: React$ComponentType<{|
     basename?: string,
@@ -155,7 +160,7 @@ declare module "react-router-dom" {
     location?: Location
   |}>
 
-  declare export function withRouter<Props: {}, Component: React$ComponentType<Props>>(
+  declare export function withRouter<Props: {...}, Component: React$ComponentType<Props>>(
     WrappedComponent: Component
   ): React$ComponentType<$Diff<React$ElementConfig<Component>, ContextRouterVoid>>;
 
@@ -163,7 +168,8 @@ declare module "react-router-dom" {
     path?: string,
     exact?: boolean,
     sensitive?: boolean,
-    strict?: boolean
+    strict?: boolean,
+    ...
   };
 
   declare export function matchPath(
@@ -172,5 +178,5 @@ declare module "react-router-dom" {
     parent?: Match
   ): null | Match;
 
-  declare export function generatePath(pattern?: string, params?: { +[string]: mixed }): string;
+  declare export function generatePath(pattern?: string, params?: { +[string]: mixed, ... }): string;
 }
