@@ -1,5 +1,5 @@
-// flow-typed signature: 458f44d0168ad45ff223678feecd6658
-// flow-typed version: 1204f9886b/reselect_v4.x.x/flow_>=v0.47.x
+// flow-typed signature: 8a7a8f67ac996c50b17930f0adc8cbe4
+// flow-typed version: c6154227d1/reselect_v4.x.x/flow_>=v0.104.x
 
 type ExtractReturnType = <Return>((...rest: any[]) => Return) => Return;
 
@@ -10,10 +10,11 @@ declare module "reselect" {
   declare type OutputSelector<-TState, TProps, TResult> =
     & InputSelector<TState, TProps, TResult>
     & {
-      recomputations(): number,
-      resetRecomputations(): number,
-      resultFunc(state: TState, props: TProps, ...rest: Array<any>): TResult,
-    };
+    recomputations(): number,
+    resetRecomputations(): number,
+    resultFunc(state: TState, props: TProps, ...rest: Array<any>): TResult,
+    ...
+  };
 
   declare type SelectorCreator = {
     <TState, TProps, TResult, T1>(
@@ -24,7 +25,6 @@ declare module "reselect" {
       selectors: [InputSelector<TState, TProps, T1>],
       resultFunc: (arg1: T1) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -34,7 +34,6 @@ declare module "reselect" {
       selectors: [InputSelector<TState, TProps, T1>, InputSelector<TState, TProps, T2>],
       resultFunc: (arg1: T1, arg2: T2) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -49,7 +48,6 @@ declare module "reselect" {
       ],
       resultFunc: (arg1: T1, arg2: T2, arg3: T3) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -66,7 +64,6 @@ declare module "reselect" {
       ],
       resultFunc: (arg1: T1, arg2: T2, arg3: T3, arg4: T4) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4, T5>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -85,7 +82,6 @@ declare module "reselect" {
       ],
       resultFunc: (arg1: T1, arg2: T2, arg3: T3, arg4: T4, arg5: T5) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -120,7 +116,6 @@ declare module "reselect" {
         arg6: T6
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -159,7 +154,6 @@ declare module "reselect" {
         arg7: T7
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -202,7 +196,6 @@ declare module "reselect" {
         arg8: T8
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -249,7 +242,6 @@ declare module "reselect" {
         arg9: T9
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -300,7 +292,6 @@ declare module "reselect" {
         arg10: T10
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <TState, TProps, TResult, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
       selector1: InputSelector<TState, TProps, T1>,
       selector2: InputSelector<TState, TProps, T2>,
@@ -355,7 +346,6 @@ declare module "reselect" {
         arg11: T11
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <
       TState,
       TProps,
@@ -446,7 +436,6 @@ declare module "reselect" {
         arg12: T12
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <
       TState,
       TProps,
@@ -543,7 +532,6 @@ declare module "reselect" {
         arg13: T13
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <
       TState,
       TProps,
@@ -646,7 +634,6 @@ declare module "reselect" {
         arg14: T14
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <
       TState,
       TProps,
@@ -755,7 +742,6 @@ declare module "reselect" {
         arg15: T15
       ) => TResult
     ): OutputSelector<TState, TProps, TResult>,
-
     <
       TState,
       TProps,
@@ -869,26 +855,25 @@ declare module "reselect" {
         arg15: T15,
         arg16: T16
       ) => TResult
-    ): OutputSelector<TState, TProps, TResult>
+    ): OutputSelector<TState, TProps, TResult>,
+    ...
   };
 
   declare type Reselect = {
     createSelector: SelectorCreator,
-
     defaultMemoize: <TFunc: Function>(
       func: TFunc,
       equalityCheck?: (a: any, b: any) => boolean
     ) => TFunc,
-
     createSelectorCreator: (
       memoize: Function,
       ...memoizeOptions: any[]
     ) => SelectorCreator,
-
-    createStructuredSelector: <TState, TProps, InputSelectors: {[k: string | number]: InputSelector<TState, TProps, any>}>(
+    createStructuredSelector: <TState, TProps, InputSelectors: { [k: string | number]: InputSelector<TState, TProps, any>, ... }>(
       inputSelectors: InputSelectors,
       selectorCreator?: SelectorCreator
-    ) => OutputSelector<TState, TProps, $ObjMap<InputSelectors, ExtractReturnType>>
+    ) => OutputSelector<TState, TProps, $ObjMap<InputSelectors, ExtractReturnType>>,
+    ...
   };
 
   declare module.exports: Reselect;
