@@ -1,5 +1,5 @@
-// flow-typed signature: ae65aae767b5a25bf5891890a98c7fc6
-// flow-typed version: aa8995a1b9/yup_v0.x.x/flow_>=v0.69.0
+// flow-typed signature: 33d1ca31f0b32e1d5ee220f97fc76fce
+// flow-typed version: c6154227d1/yup_v0.x.x/flow_>=v0.104.x
 
 declare module 'yup' {
   import type Ref from 'yup/lib/Reference';
@@ -20,26 +20,30 @@ declare module 'yup' {
     abortEarly?: boolean,
     stripUnknown?: boolean,
     recursive?: boolean,
-    context?: {},
+    context?: {...},
   |};
 
   declare export interface BaseSchema<T, GetHighType> {
-    cast(value: any, options?: { context: {} }): T;
+    cast(value: any, options?: { context: {...}, ... }): T;
     resolve(options?: any): $Call<GetHighType, T>;
     validate(value: any, options?: ValidateOptions): Promise<T>;
     validateSync(
       value: any,
-      options?: $Diff<ValidateOptions, { sync: any }>
+      options?: $Diff<ValidateOptions, { sync: any, ... }>
     ): T;
     validateAt(
       path: string,
       value: T,
-      options?: $Diff<ValidateOptions, { path: any }>
+      options?: $Diff<ValidateOptions, { path: any, ... }>
     ): Promise<T>;
     validateSyncAt(
       path: string,
       value: T,
-      options?: $Diff<ValidateOptions, { sync: any, path: any }>
+      options?: $Diff<ValidateOptions, {
+        sync: any,
+        path: any,
+        ...
+      }>
     ): T;
   }
 
@@ -54,7 +58,7 @@ declare module 'yup' {
     isValid(value: any, options?: ValidateOptions): Promise<boolean>;
     isValidSync(
       value: any,
-      options?: $Diff<ValidateOptions, { sync: any }>
+      options?: $Diff<ValidateOptions, { sync: any, ... }>
     ): boolean;
 
     isType(value: any): boolean;
@@ -106,7 +110,7 @@ declare module 'yup' {
     | string
     | null
     | void
-    | {};
+    | {...};
 
   declare export type WhenOptionsBuilderObject = {|
     is: WhenOptionsBuilderObjectIs,
@@ -131,15 +135,20 @@ declare module 'yup' {
       value: any
     ) => boolean | ValidationError | Promise<boolean | ValidationError>,
     message?: TestOptionsMessage,
-    params?: {},
+    params?: {...},
     exclusive?: boolean,
   |};
 
   declare export type SchemaDescription = {
     type: string,
     label: string,
-    meta: {},
-    tests: Array<{ name: string, params: {} }>,
+    meta: {...},
+    tests: Array<{
+      name: string,
+      params: {...},
+      ...
+    }>,
+    ...
   };
 
   declare export function ref<T>(
@@ -170,19 +179,18 @@ declare module 'yup' {
 }
 
 declare module 'yup/lib/setLocale' {
-  declare export type LocaleDescriptor<Keys> = {
-    [Keys]: string,
-  };
+  declare export type LocaleDescriptor<Keys> = { [Keys]: string, ... };
 
   declare export type LocaleObject = {|
     mixed?: {
+      ['default' | 'required' | 'oneOf' | 'notOneOf']: string,
       notType: ({|
         path: string,
         type: any,
         value: any,
         originalValue: any,
       |}) => string,
-      ['default' | 'required' | 'oneOf' | 'notOneOf']: string,
+      ...
     },
     string?: LocaleDescriptor<
       | 'length'
@@ -205,8 +213,8 @@ declare module 'yup/lib/setLocale' {
       | 'negative'
       | 'integer'
     >,
-    boolean?: {},
-    bool?: {},
+    boolean?: {...},
+    bool?: {...},
     date?: LocaleDescriptor<'min' | 'max'>,
     array?: LocaleDescriptor<'min' | 'max'>,
     object?: LocaleDescriptor<'noUnknown'>,
@@ -224,7 +232,7 @@ declare module 'yup/lib/ValidationError' {
     type: any;
     errors: Array<string>;
     inner: Array<ValidationError>;
-    params?: {};
+    params?: {...};
 
     static isError(err: any): boolean;
     static formatError(
@@ -253,8 +261,8 @@ declare module 'yup/lib/Reference' {
     +__isYupRef: true;
 
     constructor(key: string, options: RefOptions): Reference<T>;
-    getValue(options: {}): any;
-    cast(value: any, options: {}): any;
+    getValue(options: {...}): any;
+    cast(value: any, options: {...}): any;
     resolve(): Reference<T>;
     describe(): {| type: 'ref', key: string |};
 
@@ -283,14 +291,12 @@ declare module 'yup/lib/Reference' {
 declare module 'yup/lib/util/reach' {
   import type { Schema } from 'yup';
 
-  declare export default {
-    <T, GetterFn, R, ReturnGetterFn>(
-      schema: Schema<T, GetterFn>,
-      path: string,
-      value?: any,
-      context?: any
-    ): Schema<R, ReturnGetterFn>,
-  };
+  declare export default { <T, GetterFn, R, ReturnGetterFn>(
+    schema: Schema<T, GetterFn>,
+    path: string,
+    value?: any,
+    context?: any
+  ): Schema<R, ReturnGetterFn>, ... };
 }
 declare module 'yup/lib/util/isSchema' {
   declare export default (obj: any) => boolean;
@@ -348,7 +354,11 @@ declare module 'yup/lib/string' {
       regex: RegExp,
       messageOrOptions?:
         | TestOptionsMessage
-        | { message?: TestOptionsMessage, excludeEmptyString?: boolean }
+        | {
+        message?: TestOptionsMessage,
+        excludeEmptyString?: boolean,
+        ...
+      }
     ): StringSchema<T>;
 
     nullable(isNullable?: true): StringSchema<?T>;
@@ -480,8 +490,8 @@ declare module 'yup/lib/object' {
 
   declare export type ExtractSchemaType = <V>(v: BaseSchema<V, any>) => V;
 
-  declare export type ObjectSchemaConstructor = Class<ObjectSchema<{}>> &
-    (() => ObjectSchema<{}>) &
+  declare export type ObjectSchemaConstructor = Class<ObjectSchema<{...}>> &
+    (() => ObjectSchema<{...}>) &
     (<U>(obj: U) => ObjectSchema<$ObjMap<U, ExtractSchemaType>>);
 
   declare export type GetterObjectSchema = <T>(T) => ObjectSchema<T>;
@@ -497,7 +507,7 @@ declare module 'yup/lib/object' {
       fromKey: From,
       toKey: To,
       alias: true
-    ): ObjectSchema<T & { [To]: $ElementType<T, From> }>;
+    ): ObjectSchema<T & { [To]: $ElementType<T, From>, ... }>;
 
     /*
      * No ability to describe returned Schema type in this case
